+++ Apps/SpriteJoinDemo/SpriteJoinDemo/GameScene.swift
@@ -251,9 +251,9 @@ class GameScene: SKScene {
             })
         
         enterSelection
-            .attr("alpha", toValue:0)
+            .attr("alpha", toValue:CGFloat(0))
             .transition(duration:0.5)
-            .attr("alpha", toValue:1)
+            .attr("alpha", toValue:CGFloat(1))
         
     }
